Relationship data structures now follow the spec.
Previously, relationships were represented by a map[string]*Object which 
limits us to only being able to represent 1-1 relationships. In addition,
the relationship format was not adhering to the format defined by the 
spec. http://jsonapi.org/format/#document-resource-object-relationships

This commit models relationships as their own structures, and implements 
a custom Unmarshal to be able to handle 1-M relationships.