Enhanced Error Handling
Provides a StatusCode() call to the generic ErrorType. Abstracts away the need
for an end user to require type checking in order to determine the error
status code associated with the type.

Adds more documentation around errors, and adds an additional Validation check
to provide a specific error for a non-set error status.